runtime/pyproject.toml
[project]
name = "anyforce"
version = "0.16.6"
description = ""
authors = [
    {name = "exherb", email = "i@4leaf.me"},
]
dependencies = [
    "fastapi==0.75.2",
    "uvicorn[standard]==0.17.6",
    "pydantic[email]==1.9.0",
    "python-jose[cryptography]==3.3.0",
    "tortoise-orm==0.19.0",
    "passlib[bcrypt]==1.7.4",
    "requests==2.27.1",
    "types-requests==2.27.20",
    "orjson==3.6.8",
    "python-json-logger==2.0.2",
    "Jinja2==3.1.1",
    "colorama==0.4.4",
    "aiohttp[speedups]==3.8.1",
    "starlette-context==0.3.3",
    "itsdangerous==2.1.2",
    "inflection==0.5.1",
    "aenum==3.1.11",
    "python-dateutil==2.8.2",
    "rq==1.10.1",
]
requires-python = ">=3.8"
license = {text = "COPYRIGHT"}

[project.urls]
homepage = "https://github.com/42signal/anyforce"


[project.optional-dependencies]
[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"

[tool]
[tool.pdm]
includes = [
    "anyforce/**/*.py",
    "anyforce/py.typed",
    "README.md",
]
[[tool.pdm.source]]
name = "pypi"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
verify_ssl = true

[tool.pdm.dev-dependencies]
dev = [
    "flake8==4.0.1",
    "isort==5.10.1",
    "black==22.3.0",
    "pytest==7.1.2",
    "typer==0.4.1",
    "pytest-asyncio==0.18.3",
    "Faker==13.4.0",
    "pytest-cov==3.0.0",
    "twine==4.0.0",
]

[tool.isort]
profile = "black"
line_length = 79
multi_line_output = 3

[tool.pytest.ini_options]
testpaths = [
    "tests",
]
